Improve contributors request handling and code reuse
